RCS_Server/requirements.txt
django
channels["daphne"]
channels_redis
django-channels-redis